+++ src/components/MarketPriceTable/CryptoCard.jsx
@@ -1,5 +1,5 @@
 import defaultImg from '/src/assets/no-image.jpg';
-import styles from '/src/components/MarketPriceTable/MarketPriceTable.css';
+import styles from '/src/components/MarketPriceTable/MarketPriceTable.module.css';
 import classNames from 'classnames/bind';
 
 const cn = classNames.bind(styles);
