+++ src/components/MarketPriceTable/MarketPriceTable.module.css
