+++ src/components/MarketPriceTable/MarketPriceTable_api.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, useRef } from 'react';
 import { useCurrency } from '/src/contexts/CurrencyContext';
 import { DataGrid } from '@mui/x-data-grid';
-import styles from '/src/components/MarketPriceTable/MarketPriceTable.css';
+import styles from '/src/components/MarketPriceTable/MarketPriceTable.module.css';
 import classNames from 'classnames/bind';
 import CustomPagination from '/src/components/MarketPriceTable/CustomPagintion';
 import fetchMarketData from '/src/components/MarketPriceTable/api/marketDataApi';
