+++ src/components/MarketPriceTable/PriceChangePercentage.jsx
@@ -1,5 +1,5 @@
 import shortenNumber from '/src/components/MarketPriceTable/utils/shortenNumber.js';
-import styles from '/src/components/MarketPriceTable/MarketPriceTable.css';
+import styles from '/src/components/MarketPriceTable/MarketPriceTable.module.css';
 import classNames from 'classnames/bind';
 
 const cn = classNames.bind(styles);
