+++ src/components/MarketPriceTable/columns.jsx
@@ -3,7 +3,7 @@ import PriceChangePercentage from '/src/components/MarketPriceTable/PriceChangeP
 import CoinTotalVolume from '/src/components/MarketPriceTable/CoinTotalVolume';
 import CryptoCard from '/src/components/MarketPriceTable/CryptoCard';
 import formatPrice from '/src/components/MarketPriceTable/utils/formatPrice.js';
-import styles from '/src/components/MarketPriceTable/MarketPriceTable.css';
+import styles from '/src/components/MarketPriceTable/MarketPriceTable.module.css';
 import classNames from 'classnames/bind';
 
 const cn = classNames.bind(styles);
